Move verbose options out of parser.ts
