+++ src/config.ts
@@ -1,0 +1,2 @@
+export const VERBOSE = 1; // 0 none, 1 some, 2 a lot, 3 everything
+export const WARNINGS = 1; // 0 none, 1 some, 2 a lot, 3 everything
