+++ src/parse/index.ts
@@ -3,7 +3,7 @@ import AS3Parser from './parser';
 import SourceFile from './source-file';
 import AS3Scanner from './scanner';
 import {parseCompilationUnit} from './parse-declarations';
-import {VERBOSE, WARNINGS} from './parser';
+import {VERBOSE, WARNINGS} from '../config';
 
 export default function parse(filePath:string, content:string):Node {
