+++ src/parse/parser.ts
@@ -41,9 +41,6 @@ export const NEW_LINE = '\n';
 const SINGLE_LINE_COMMENT = '//';
 export const VECTOR = 'Vector';
 
-export const VERBOSE = 0; // 0 none, 1 some, 2 a lot, 3 everything
-export const WARNINGS = 1; // 0 none, 1 some, 2 a lot, 3 everything
-
 /**
  * @author xagnetti
  */
